+++ frontend/src/views/BoardView/BoardView.styles.ts
@@ -37,7 +37,7 @@ export const NavLinkButton = styled(Link)<NavLinkProps>`
 
 export const BoardGrid = styled.div`
   display: grid;
-  grid-template-columns: repeat(4, 1fr);
+  grid-template-columns: repeat(5, 1fr);
   gap: 20px;
   overflow-x: auto;
   align-items: start;